requirements.txt
flask
pymongo
dnspython
requests
python-dotenv
pdfkit
